requirements.txt
matplotlib==3.8.0
matplotlib-inline==0.1.6
numpy==1.26.1
oauthlib==3.2.0
PyGithub==1.55
pytest==6.2.5
pytest-cov==3.0.0
PyYAML==6.0.1
scipy==1.11.3
tqdm==4.62.3
ipython==8.19.0
tensorflow==2.15.0.post1
gymnasium==0.29.1
gymnasium[classic-control]
gymnasium[mujoco]
tensorflow-addons==0.23.0
hydra-core==1.3.2
mujoco==3.1.1